+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 - Add script runtime function to update groups.
 - Add script runtime function to list groups a user is part of.
 - Add script runtime function to list users belonging to a group.
+- Send in-app notification on friend request.
+- Send in-app notification on friend request accept.
+- Send in-app notification when a Facebook friend signs into the game for the first time.
+- Send in-app notification to group admins when a user requests to join a private group.
+- Send in-app notification to the user when they are added to a group or their request to join a private group is accepted.
+- Send in-app notification to the user when someone wants to DM chat.
 
 ### Changed
 - Use Lua table for Content field when creating new notifications.